backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.5
SQLAlchemy==2.0.32
python-dotenv==1.0.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
# OCR dependencies
pillow==10.4.0
pytesseract==0.3.10
pdf2image==1.17.0
python-multipart==0.0.9
# Security
python-jose==3.3.0
cryptography==42.0.5
requests==2.32.3
python-magic==0.4.26
# pyodbc==5.1.0